test find path

+++ src/node/node.py
@@ -1,6 +1,6 @@
 
 class Node:
-    def __init__(self, node_id: int, x: int, y: int) -> None:
+    def __init__(self, node_id: int, x: float, y: float) -> None:
         self.node_id = node_id
         self.x = x
         self.y = y
